+++ apps/wallet-mobile/src/features/Discover/useCases/BrowseDapp/BrowseDappScreen.tsx
