+++ apps/wallet-mobile/src/features/Discover/useCases/BrowseDapp/BrowserSearchToolbar.tsx
