+++ apps/wallet-mobile/src/features/Discover/useCases/BrowseDapp/BrowserTabBar.tsx
