+++ apps/wallet-mobile/src/features/Discover/useCases/BrowseDapp/BrowserTabsBar.tsx
