+++ apps/wallet-mobile/src/features/Discover/useCases/BrowseDapp/BrowserToolbar.tsx
