+++ apps/wallet-mobile/src/features/Discover/useCases/SearchDappInBrowser/SearchDappInBrowserScreen.stories.tsx
@@ -4,16 +4,16 @@ import * as React from 'react'
 import {SelectedWalletProvider} from '../../../../features/WalletManager/Context'
 import {mocks as walletMocks} from '../../../../yoroi-wallets/mocks'
 import {BrowserProvider} from '../../common/BrowserProvider'
-import {BrowseSearchDappScreen} from './BrowseSearchDappScreen'
+import {SearchDappInBrowserScreen} from './SearchDappInBrowserScreen'
 
-storiesOf('Discover BrowseSearchDappScreen', module)
+storiesOf('Discover SearchDappInBrowserScreen', module)
   .addDecorator((story) => <SelectedWalletProvider wallet={walletMocks.wallet}>{story()}</SelectedWalletProvider>)
   .add('initial', () => <Initial />)
 
 const Initial = () => {
   return (
     <BrowserProvider>
-      <BrowseSearchDappScreen />
+      <SearchDappInBrowserScreen />
     </BrowserProvider>
   )
 }
