+++ apps/wallet-mobile/src/features/Discover/useCases/SelectDappFromList/DAppItem/DAppItem.tsx
