+++ apps/wallet-mobile/src/features/Discover/useCases/SelectDappFromList/SelectDappFromListScreen.stories.tsx
@@ -5,9 +5,9 @@ import {SelectedWalletProvider} from '../../../../features/WalletManager/Context
 import {SearchProvider} from '../../../../Search/SearchContext'
 import {mocks as walletMocks} from '../../../../yoroi-wallets/mocks'
 import {BrowserProvider} from '../../common/BrowserProvider'
-import {DiscoverListScreen} from './DiscoverListScreen'
+import {SelectDappFromListScreen} from './SelectDappFromListScreen'
 
-storiesOf('Discover DiscoverListScreen', module)
+storiesOf('Discover SelectDappFromListScreen', module)
   .addDecorator((story) => <SelectedWalletProvider wallet={walletMocks.wallet}>{story()}</SelectedWalletProvider>)
   .add('initial', () => <Initial />)
   .add('searching dapp', () => <SearchDApp />)
@@ -16,7 +16,7 @@ const Initial = () => {
   return (
     <SearchProvider>
       <BrowserProvider>
-        <DiscoverListScreen />
+        <SelectDappFromListScreen />
       </BrowserProvider>
     </SearchProvider>
   )
@@ -31,7 +31,7 @@ const SearchDApp = () => {
       }}
     >
       <BrowserProvider>
-        <DiscoverListScreen />
+        <SelectDappFromListScreen />
       </BrowserProvider>
     </SearchProvider>
   )
